@@ -5,21 +5,18 @@ import 'package:help_mate/session/tagSession.dart';
 import 'package:help_mate/testControllerForAnuj.dart';
 
 class TestForAnuj extends StatefulWidget {
-  const TestForAnuj({Key key}) : super(key: key);
+  const TestForAnuj({Key? key}) : super(key: key);
 
   @override
   _TestForAnujState createState() => _TestForAnujState();
 }
 
 class _TestForAnujState extends State<TestForAnuj> {
-
-
   TestControllerForAnuj controller = TestControllerForAnuj();
 
   TagSession session = TagSession();
 
-  List tagsList=[];
-
+  List tagsList = [];
 
   @override
   void initState() {
@@ -29,46 +26,42 @@ class _TestForAnujState extends State<TestForAnuj> {
     fetchTags();
   }
 
-
-  fetchTags() async{
+  fetchTags() async {
     dynamic data = await session.getAllTags();
 
-    if(data == null)
-      {
-        print("Unable to get data");
-      }
-    else{
-    setState(() {
-      tagsList=data;
-    });
+    if (data == null) {
+      print("Unable to get data");
+    } else {
+      setState(() {
+        tagsList = data;
+      });
     }
   }
-
-
 
   @override
   Widget build(BuildContext context) {
     final items = List<String>.generate(10000, (i) => "Item $i");
     return Scaffold(
-      appBar: AppBar(title: Text("Hey"),),
-      body:  Container(
+      appBar: AppBar(
+        title: Text("Hey"),
+      ),
+      body: Container(
         color: Colors.grey[500],
         child: Column(
-
           children: [
-        Expanded(
-          child: ListView.builder(
-          padding: const EdgeInsets.all(8),
-            itemCount: tagsList.length,
-            itemBuilder: (BuildContext context, int index) {
-              return Container(
-                height: 50,
-                child: Center(child: Text('Entry ${tagsList[index]['tagDescription']}')),
-              );
-            }
-      ),
-        ),
-
+            Expanded(
+              child: ListView.builder(
+                  padding: const EdgeInsets.all(8),
+                  itemCount: tagsList.length,
+                  itemBuilder: (BuildContext context, int index) {
+                    return Container(
+                      height: 50,
+                      child: Center(
+                          child: Text(
+                              'Entry ${tagsList[index]['tagDescription']}')),
+                    );
+                  }),
+            ),
             OutlinedButton(
               onPressed: () {
                 print('Received click');
@@ -81,9 +74,9 @@ class _TestForAnujState extends State<TestForAnuj> {
                 print('Received click');
                 controller.updateTagData();
               },
-              child: const Text('Click me to Update Tag Data for 0qnqZleV73ZMcNeDZ6fG id'),
+              child: const Text(
+                  'Click me to Update Tag Data for 0qnqZleV73ZMcNeDZ6fG id'),
             ),
-
           ],
         ),
       ),
