@@ -1,42 +1,39 @@
 class CommentData {
-  String commentId, postId,userId,commentContent;
-  int timeStamp,likesOnComment;
-  List<String> tagIds, whoLiked;
-
+  final String commentId, postId, userId, commentContent;
+  final int timeStamp, likesOnComment;
+  final List<String> tagIds, whoLiked;
 
   CommentData.name(
-      this.commentId,
-      this.postId,
-      this.userId,
-      this.commentContent,
-      this.timeStamp,
-      this.likesOnComment,
-      this.tagIds,
-      this.whoLiked);
+      {required this.commentId,
+      required this.postId,
+      required this.userId,
+      required this.commentContent,
+      required this.timeStamp,
+      required this.likesOnComment,
+      required this.tagIds,
+      required this.whoLiked});
 
-  CommentData.fromJson(Map<String, dynamic> json) {
-    commentId= json['commentId'];
-    postId= json['postId'];
-    userId= json['userId'];
-    commentContent= json['commentContent'];
-    timeStamp= json['timeStamp'];
-    likesOnComment=json['likesOnComment'];
-    tagIds = json['tagIds'];
-    whoLiked=json['whoLiked'];
-  }
+  CommentData.fromJson(Map<String, dynamic> json)
+      : commentId = json['commentId'],
+        postId = json['postId'],
+        userId = json['userId'],
+        commentContent = json['commentContent'],
+        timeStamp = json['timeStamp'],
+        likesOnComment = json['likesOnComment'],
+        tagIds = json['tagIds'],
+        whoLiked = json['whoLiked'];
 
   Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
+    final Map<String, dynamic> data = {};
     data['commentId'] = this.commentId;
     data['postId'] = this.postId;
-    data['userId']= this.userId;
-    data['commentContent']= this.commentContent;
-    data['timeStamp']= this.timeStamp;
+    data['userId'] = this.userId;
+    data['commentContent'] = this.commentContent;
+    data['timeStamp'] = this.timeStamp;
     data['likesOnComment'] = this.likesOnComment;
-    data['tagIds']= this.tagIds;
-    data['whoLiked']= this.whoLiked;
+    data['tagIds'] = this.tagIds;
+    data['whoLiked'] = this.whoLiked;
 
     return data;
   }
-
 }
