@@ -1,21 +1,20 @@
 class TagData {
+  final String tagName, tagDescription, tagId;
 
-  String tagId,tagName,tagDescription;
+  TagData(
+      {required this.tagName,
+      required this.tagDescription,
+      required this.tagId});
 
-
-  TagData(this.tagName, this.tagDescription);
-
-
-  TagData.fromJson(Map<String, dynamic> json) {
-    tagId= json['tagId'];
-    tagName= json['tagName'];
-    tagDescription= json['tagDescription'];
-  }
+  TagData.fromJson(Map<String, dynamic> json)
+      : tagId = json['tagId'],
+        tagDescription = json['tagDescription'],
+        tagName = json['tagName'];
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     data['tagName'] = this.tagName;
-    data['tagDescription']= this.tagDescription;
+    data['tagDescription'] = this.tagDescription;
     return data;
   }
 }
