@@ -1,10 +1,9 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:help_mate/modelClasses/commentData.dart';
 
-class CommentSession{
+class CommentSession {
   CollectionReference commentCollection =
-  FirebaseFirestore.instance.collection("commentCollection");
-
+      FirebaseFirestore.instance.collection("commentCollection");
 
   Stream<List<CommentData>> tagIdStram(String tagId) {
     Stream<QuerySnapshot> snapshot = commentCollection
@@ -12,11 +11,10 @@ class CommentSession{
         .limit(20)
         .snapshots();
 
-    return snapshot.map((qSnap) =>
-        qSnap.docs.map((doc) => CommentData.fromJson(doc.data())).toList());
+    return snapshot.map((qSnap) => qSnap.docs
+        .map((doc) => CommentData.fromJson(doc.data() as Map<String, dynamic>))
+        .toList());
   }
-
-
 
   Future<String> pushNewComment(CommentData commentData) async {
     Map<String, dynamic> commentItemData = commentData.toJson();
@@ -27,14 +25,10 @@ class CommentSession{
     return doc.id.toString();
   }
 
-
-  Future pushUpdateComment(CommentData commentData,String commentId) async
-  {
+  Future pushUpdateComment(CommentData commentData, String commentId) async {
     Map<String, dynamic> data = commentData.toJson();
     commentCollection.doc(commentId).update(data);
   }
-
-
 
   Future getComment(String commentId) async {
     commentCollection
@@ -49,6 +43,4 @@ class CommentSession{
       }
     });
   }
-
-
 }
